+++ src/web/assets/cp/src/js/AssetIndex.js
@@ -19,7 +19,6 @@ Craft.AssetIndex = Craft.BaseElementIndex.extend({
 
   _uploadTotalFiles: 0,
   _uploadFileProgress: {},
-  _uploadedAssetIds: [],
   _currentUploaderSettings: {},
 
   _assetDrag: null,
@@ -960,7 +959,7 @@ Craft.AssetIndex = Craft.BaseElementIndex.extend({
 
     if (response.success || response.conflict) {
       // Add the uploaded file to the selected ones, if appropriate
-      this._uploadedAssetIds.push(response.assetId);
+      this.selectElementAfterUpdate(response.assetId);
 
       // If there is a prompt, add it to the queue
       if (response.conflict) {
@@ -1009,7 +1008,8 @@ Craft.AssetIndex = Craft.BaseElementIndex.extend({
    */
   _updateAfterUpload: function () {
     if (this.settings.context !== 'index') {
-      this.setSortAttribute('dateModified');
+      this.clearSearch();
+      this.setSortAttribute('dateCreated');
       this.setSortDirection('desc');
     }
     this.updateElements();
@@ -1041,7 +1041,7 @@ Craft.AssetIndex = Craft.BaseElementIndex.extend({
 
       var followupCallback = (data, textStatus) => {
         if (textStatus === 'success' && data.assetId) {
-          this._uploadedAssetIds.push(data.assetId);
+          this.selectElementAfterUpdate(data.assetId);
         } else if (data.error) {
           alert(data.error);
         }
@@ -1109,18 +1109,6 @@ Craft.AssetIndex = Craft.BaseElementIndex.extend({
       }
 
       this._assetDrag.addItems($newElements.has('div.element[data-movable]'));
-    }
-
-    // See if we have freshly uploaded files to add to selection
-    if (this._uploadedAssetIds.length) {
-      if (this.view.settings.selectable) {
-        for (var i = 0; i < this._uploadedAssetIds.length; i++) {
-          this.view.selectElementById(this._uploadedAssetIds[i]);
-        }
-      }
-
-      // Reset the list.
-      this._uploadedAssetIds = [];
     }
 
     this.base(append, $newElements);